feature:finance

+++ inventory/models.py
@@ -40,13 +40,13 @@ class Product(models.Model):
     ]
     
     name = models.CharField(max_length=255)
-    quantity = models.IntegerField()
+    quantity = models.FloatField()
     cost = models.DecimalField(max_digits=10, decimal_places=2)
     price = models.DecimalField(max_digits=10, decimal_places=2, default=1) 
     unit = models.ForeignKey(UnitOfMeasurement, on_delete=models.CASCADE, null=True)
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
     tax_type = models.CharField(max_length=50, choices=tax_choices)
-    min_stock_level = models.IntegerField(default=0, null=True)
+    min_stock_level = models.FloatField(default=0, null=True)
     raw_material = models.BooleanField(default=False)
     finished_product = models.BooleanField(default=False)
     description = models.TextField()
@@ -56,7 +56,7 @@ class Product(models.Model):
         return self.name
 class ProductionRawMaterials(models.Model):
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    quantity = models.IntegerField()
+    quantity = models.FloatField()
     
     def __str__(self) -> str:
         return self.product.name
@@ -91,10 +91,10 @@ class ProductionItems(models.Model):
     declared = models.BooleanField(default=False)
     left_overs = models.FloatField(default=0, null=True, blank=True)
     wastage = models.FloatField(default=0, null=True, blank=True)      
-    portions = models.IntegerField(default=0, null=True)
-    staff_portions = models.IntegerField(default=0, null=True)  
-    declared_quantity = models.IntegerField(default=0, null=True)
-    portions_sold = models.IntegerField(default=0, null=True)
+    portions = models.FloatField(default=0, null=True)
+    staff_portions = models.FloatField(default=0, null=True)  
+    declared_quantity = models.FloatField(default=0, null=True)
+    portions_sold = models.FloatField(default=0, null=True)
 
 class MinorProductionItems(models.Model):
     production = models.ForeignKey(Production, on_delete=models.CASCADE)
@@ -202,9 +202,9 @@ class PurchaseOrderItem(models.Model):
 
     purchase_order = models.ForeignKey(PurchaseOrder, on_delete=models.CASCADE, related_name='items')
     product = models.ForeignKey(Product, on_delete=models.SET_NULL, null=True)
-    quantity = models.IntegerField()
+    quantity = models.FloatField()
     unit_cost = models.DecimalField(max_digits=10, decimal_places=2)
-    received_quantity = models.IntegerField(default=0) 
+    received_quantity = models.FloatField(default=0) 
     received = models.BooleanField(default=False, null=True)
     note = models.CharField(default='', null=True, max_length=255)
 
@@ -255,8 +255,8 @@ class Logs(models.Model):
     user = models.ForeignKey('users.User', on_delete=models.SET_NULL, null=True)
     action = models.CharField(max_length=20, choices=ACTION_CHOICES)
     purchase_order = models.ForeignKey(PurchaseOrder, null=True, blank=True, on_delete=models.SET_NULL)
-    quantity = models.IntegerField()
-    total_quantity = models.IntegerField()
+    quantity = models.FloatField()
+    total_quantity = models.FloatField()
     timestamp = models.DateField(auto_now_add=True)
     description = models.CharField(max_length=255, null=True)
